requirements.txt
amqp==2.5.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==1.10.6
django-celery==3.2.2
django-celery-results==1.0.4
idna==2.8
kombu==4.2.2.post1
psycopg2==2.7.6.1
pyactiveresource==2.1.2
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
ShopifyAPI==5.0.1
six==1.12.0
urllib3==1.25.3
vine==1.3.0